Better Parsing of Search Params
